Refine the Sidebar
