+++ src/modules/main/types/mainTypes.ts
@@ -1,0 +1,7 @@
+export interface ItemTypes {
+    title: string;
+    to: string;
+    icon: React.ReactNode;
+    selected: string;
+    handleItemClick: any;
+  }
